refactoring: move produceexp from unit.c to study.c
