+++ src/kernel/unit.c
@@ -28,7 +28,6 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include "curse.h"
 #include "item.h"
 #include "move.h"
-#include "monster.h"
 #include "order.h"
 #include "plane.h"
 #include "race.h"
@@ -1890,25 +1889,6 @@ bool unit_name_equals_race(const unit *u) {
 
 bool unit_can_study(const unit *u) {
     return !((u_race(u)->flags & RCF_NOLEARN) || fval(u, UFL_WERE));
-}
-
-static double produceexp_chance(void) {
-    return config_get_flt("study.from_use", 1.0 / 3);
-}
-
-void produceexp_ex(struct unit *u, skill_t sk, int n, bool(*learn)(unit *, skill_t, double))
-{
-    if (n != 0 && (is_monsters(u->faction) || playerrace(u_race(u)))) {
-        double chance = produceexp_chance();
-        if (chance > 0.0F) {
-            learn(u, sk, (n * chance) / u->number);
-        }
-    }
-}
-
-void produceexp(struct unit *u, skill_t sk, int n)
-{
-    produceexp_ex(u, sk, n, learn_skill);
 }
 
 /* ID's für Einheiten und Zauber */
